# GreekLearningApp-Frontend/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Greek Parser API",
    "description": "API for Greek language learning and parsing",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5001"
    }
  ],
  "paths": {
...
    "/api/grammaticalForms": {
      "get": {
        "tags": [
          "GrammaticalFeatures"
        ],
        "operationId": "GrammaticalFeatures_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimpleGrammaticalFormDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "GrammaticalFeatures"
        ],
        "operationId": "GrammaticalFeatures_Create",
        "requestBody": {
          "x-name": "createDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimpleGrammaticalFormDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleGrammaticalFormDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/grammaticalForms/{id}": {
      "get": {
        "tags": [
          "GrammaticalFeatures"
        ],
        "operationId": "GrammaticalFeatures_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleGrammaticalFormDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "operationId": "Lessons_GetAllLessons",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lessons"
        ],
        "operationId": "Lessons_CreateLesson",
        "requestBody": {
          "x-name": "lessonDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{id}": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "operationId": "Lessons_GetLesson",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{id}/complete": {
      "post": {
        "tags": [
          "Lessons"
        ],
        "operationId": "Lessons_CompleteLesson",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "completionDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteLessonDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Create",
        "requestBody": {
          "x-name": "createDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "updateDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SimpleGrammaticalFormDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "grammarGUID": {
            "type": "string",
            "format": "guid"
          },
          "grammarId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "lessonId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LessonDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "lessonIndex": {
            "type": "integer",
            "format": "int32"
          },
          "contentMarkdown": {
            "type": "string"
          },
          "lessonType": {
            "type": "string"
          },
          "grammaticalFeatureIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "syntacticalFeatureIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "vocabularyIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "isCompleted": {
            "type": "boolean"
          }
        }
      },
      "CreateLessonDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string"
          },
          "lessonIndex": {
            "type": "integer",
            "format": "int32"
          },
          "contentMarkdown": {
            "type": "string"
          },
          "lessonType": {
            "type": "string"
          },
          "grammaticalFeatureIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "syntacticalFeatureIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "vocabularyIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "CompleteLessonDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "lessonId": {
            "type": "integer",
            "format": "int32"
          },
          "score": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
...
      "CreateUserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateUserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}