# backend/Koine.Application/DTOs/Lessons/CompleteLessonDto.cs
// Koine.Application/DTOs/Lessons/CompleteLessonDto.cs
namespace Koine.Application.DTOs.Lessons
{
    public class CompleteLessonDto
    {
        public int LessonId { get; set; }
        public int? Score { get; set; }
    }
}