# backend/Koine.Application/DTOs/Users/CreateUserDto.cs
// Koine.Application/DTOs/Users/CreateUserDto.cs
namespace Koine.Application.DTOs.Users
{
    public class CreateUserDto
    {
        public string? Email { get; set; }
        public string? Username { get; set; }
        public string? Password { get; set; }
    }
}